Iyzipay added

+++ Server/Controllers/OrderController.cs
@@ -1,4 +1,7 @@
 using System.Security.Claims;
+using Iyzipay;
+using Iyzipay.Model;
+using Iyzipay.Request;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Data.SqlClient;
@@ -12,21 +15,39 @@ namespace Server.Controller;
 [Authorize]
 [ApiController]
 [Route("api/[controller]")]
-public class OrderController(DataContext context) : ControllerBase
+public class OrderController(DataContext context, IConfiguration config) : ControllerBase
 {
     private readonly DataContext _context = context;
+    private readonly IConfiguration _config = config;
 
     [HttpGet]
-    public async Task<ActionResult<List<OrderDTO>>> GetOrders()
+    public async Task<ActionResult<PagedDataDTO<OrderDTO>>> GetOrders(int pageSize, int firstItemIndex = 0)
     {
         var customerId = User.FindFirstValue(ClaimTypes.NameIdentifier);
         if (customerId == null) return BadRequest("User not found.");
-        var order = await _context.Orders.FromSqlRaw(
-            "SELECT * FROM Orders WHERE CustomerId = @customerId", new SqlParameter("@customerId", customerId)
-            ).Include(i => i.OrderItems)
+
+        var query = _context.Orders
+            .AsQueryable();
+
+        var totalRecords = await query.CountAsync();
+
+        var orders = await query
+            .Where(i => i.CustomerId == customerId)
+            .OrderByDescending(i => i.OrderDate)
+            .Skip(firstItemIndex)
+            .Take(pageSize)
+            .Include(i => i.OrderItems)
             .Select(i => OrderToDTO(i))
             .ToListAsync();
-        return Ok(order);
+        
+        var pagedData = new PagedDataDTO<OrderDTO>{
+            Data = orders,
+            TotalRecords = totalRecords,
+            LastItemIndex = firstItemIndex + pageSize - 1,
+            PageSize = pageSize,
+            TotalPages = (int)Math.Ceiling(totalRecords / (double)pageSize)
+        };
+        return Ok(pagedData);
     }
 
     [HttpGet("{orderId}")]
@@ -63,10 +84,11 @@ public class OrderController(DataContext context) : ControllerBase
             DeliveryFee = deliveryFee,
             AddresLine = orderRequest.AddresLine,
             City = orderRequest.City,
-            FullName = orderRequest.FullName,
+            Name = orderRequest.Name,
+            Surname = orderRequest.Surname,
             Phone = orderRequest.Phone,
             SubTotal = subTotal,
-            OrderItems = cart.CartItems.Select(i => i.Product == null ? null : new OrderItem
+            OrderItems = cart.CartItems.Select(i => i.Product == null ? null : new Entity.OrderItem
             {
                 Price = i.Product.Price,
                 Quantity = i.Quantity,
@@ -74,13 +96,93 @@ public class OrderController(DataContext context) : ControllerBase
                 ProductImageUrl = i.Product.ImageUrl,
                 ProductId = i.ProductId
             }
-            ).OfType<OrderItem>().ToList()
+            ).OfType<Entity.OrderItem>().ToList()
         };
+
+        var paymentResult = await ProcessPayment(cart, orderRequest);
+        if (paymentResult.Status == "failure") return BadRequest(paymentResult.ErrorMessage);
+
+        order.ConversationId = paymentResult.ConversationId;
+        order.BasketId = paymentResult.BasketId;
+
         _context.Orders.Add(order);
         _context.Carts.Remove(cart);
+
         var success = await _context.SaveChangesAsync() > 0;
-        if (success) return CreatedAtAction(nameof(GetOrder), new {orderId= order.OrderId}, OrderToDTO(order));
+        if (success) return CreatedAtAction(nameof(GetOrder), new { orderId = order.OrderId }, OrderToDTO(order));
         return BadRequest(new ProblemDetails { Title = "The order could not be created.", Status = 400 });
+    }
+
+    private async Task<Payment> ProcessPayment(Cart cart, CreateOrderDTO orderRequest)
+    {
+        Options options = new Options();
+        options.ApiKey = _config["PaymentAPI:APIKey"];
+        options.SecretKey = _config["PaymentAPI:SecretKey"];
+        options.BaseUrl = "https://sandbox-api.iyzipay.com";
+
+        CreatePaymentRequest request = new CreatePaymentRequest();
+        request.Locale = Locale.TR.ToString();
+        request.ConversationId = Guid.NewGuid().ToString();
+        request.Price = cart.CalculateTotalCost().ToString();
+        request.PaidPrice = (cart.CalculateTotalCost() * 1.2).ToString();
+        request.Currency = Currency.TRY.ToString();
+        request.Installment = 1;
+        request.BasketId = cart.CartId.ToString();
+        request.PaymentChannel = PaymentChannel.WEB.ToString();
+        request.PaymentGroup = PaymentGroup.PRODUCT.ToString();
+
+        PaymentCard paymentCard = new PaymentCard();
+        paymentCard.CardHolderName = orderRequest.CardHolderName;
+        paymentCard.CardNumber = orderRequest.CardNumber;
+        paymentCard.ExpireMonth = orderRequest.CardExpireMonth;
+        paymentCard.ExpireYear = orderRequest.CardExpireYear;
+        paymentCard.Cvc = orderRequest.CardCvc;
+        paymentCard.RegisterCard = 0;
+        request.PaymentCard = paymentCard;
+
+        Buyer buyer = new Buyer();
+        buyer.Id = "BY789";
+        buyer.Name = orderRequest.Name;
+        buyer.Surname = orderRequest.Surname;
+        buyer.GsmNumber = orderRequest.Phone;
+        buyer.Email = orderRequest.Email;
+        buyer.IdentityNumber = User.FindFirstValue(ClaimTypes.NameIdentifier);
+        buyer.LastLoginDate = "2015-10-05 12:43:35";
+        buyer.RegistrationDate = "2013-04-21 15:12:09";
+        buyer.RegistrationAddress = orderRequest.AddresLine;
+        buyer.Ip = "85.34.78.112";
+        buyer.City = orderRequest.City;
+        buyer.Country = "Turkey";
+        buyer.ZipCode = "34732";
+        request.Buyer = buyer;
+
+        Address shippingAddress = new Address();
+        shippingAddress.ContactName = orderRequest.Name + " " + orderRequest.Surname;
+        shippingAddress.City = orderRequest.City;
+        shippingAddress.Country = "Turkey";
+        shippingAddress.Description = orderRequest.AddresLine;
+        shippingAddress.ZipCode = "34742";
+        request.ShippingAddress = shippingAddress;
+
+        request.BillingAddress = shippingAddress;
+
+        List<BasketItem> basketItems = new List<BasketItem>();
+
+
+        foreach (var item in cart.CartItems)
+        {
+            BasketItem basketItem = new BasketItem();
+            basketItem.Id = item.CartItemId.ToString();
+            basketItem.Name = item.Product.Name;
+            basketItem.Category1 = "testCategory";
+            basketItem.ItemType = BasketItemType.PHYSICAL.ToString();
+            basketItem.Price = ((double)item.Product.Price * item.Quantity).ToString();
+            basketItems.Add(basketItem);
+        }
+
+        request.BasketItems = basketItems;
+
+        return await Payment.Create(request, options);
     }
 
     private static OrderDTO OrderToDTO(Order order)
@@ -92,7 +194,8 @@ public class OrderController(DataContext context) : ControllerBase
             City = order.City,
             CustomerId = order.CustomerId,
             DeliveryFee = order.DeliveryFee,
-            FullName = order.FullName,
+            Name = order.Name,
+            Surname = order.Surname,
             OrderStatus = order.OrderStatus,
             OrderDate = order.OrderDate,
             Phone = order.Phone,
